Move IGPUBuffer interface to new Velaptor.OpenGL.Buffers namespace & folder
