+++ Velaptor/OpenGL/Buffers/IGPUBuffer.cs
@@ -2,7 +2,7 @@
 // Copyright (c) KinsonDigital. All rights reserved.
 // </copyright>
 
-namespace Velaptor.OpenGL
+namespace Velaptor.OpenGL.Buffers
 {
     /// <summary>
     /// Manages buffer data in the GPU.
